MAINT: new tools and utilities files definition
